Remove documentation of non-existing Expose.attributeName
The attribute name is hardcoded to "exposedThrough"; nothing was ever
assigned to the documented instance variable.

Since the `@renderer` decorator is the only user of the `Expose` class,
I don't think it's worthwhile to have a customizable attribute name,
so I chose to remove the `ivar` docstring field rather than implement
the feature.