Add type annotations for twisted.web.template
Since a lot of functions consume or produce "Stan objects" or "objects
that can be flattened" and a lot of types can be flattened, I added
a type alias for that: `Flattenable`. This may be useful for projects
using Twisted as well, so I exported it as a public name from
`twisted.web.template`.

For the `XMLFile` path, the annotation does not include types for
which support is deprecated. I assume projects type checking their
code do not want to use a feature that was deprecated 9 years ago.

The documentation for the `TagLoader` class used to state that it
could load `IRenderable` providers, but the implementation actually
accepts any flattenable object. The instance variable in which
the to-be-loaded object is stored is even named `tag` while the
`Tag` class does not implement `IRenderable`. Therefore I assume
the docstring was incorrect and changed it to accept `Flattenable`
instead. Note that `IRenderable` is a member of `Flattenable`, so
the previously documented interface is supported by the annotation.

I removed `@type` and `@rtype` docstring fields that provided no
additional value over the type annotations. In quite a few cases,
the documented types were incorrect, so this will increase the
accuracy of the documentation.